requirements.txt
antiorm==1.2.1
argon2-cffi==20.1.0
asgiref==3.2.9
astroid==2.4.2
attrs==20.1.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
configparser==5.0.0
crayons==0.3.1
cycler==0.10.0
cymem==2.0.3
DateTime==4.3
db==0.1.1
db-sqlite3==0.0.1
decorator==4.4.2
defusedxml==0.6.0
demjson==2.2.4
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-bootstrap4==2.1.1
django-heroku==0.3.1
django-imagekit==4.0.2
entrypoints==0.3
fastai @ git+https://github.com/fastai/fastai.git@54a9e3cf4fd0fa11fc2453a5389cc9263f6f0d77
fastprogress==0.2.4
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
inflection==0.5.1
install==1.3.3
ipykernel==5.3.2
ipython==7.16.1
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
json5==0.9.5
jsonpatch==1.26
jsonpointer==2.0
jsonschema==3.2.0
jtutils==0.0.6
jupyter-client==6.1.5
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
leven==1.0.4
lxml==4.5.2
mariadb==0.9.59
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.5.0
multitasking==0.0.9
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
nose==1.3.7
notebook==6.1.3
numexpr==2.7.1
numpy==1.19.0
nvidia-ml-py3==7.352.0
opencv-python==4.3.0.36
packaging==20.4
pandas==1.0.5
pandas-datareader==0.9.0
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==7.1.2
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.5
psycopg2==2.7.5
psycopg2-binary==2.8.5
pycparser==2.20
Pygments==2.6.1
pykrx==0.1.26
pylint==2.5.3
pyparsing==2.4.7
pyrsistent==0.16.0
python-csv==0.0.11
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
Quandl==3.5.2
requests==2.24.0
scipy==1.5.1
seaborn==0.10.1
selenium==3.141.0
Send2Trash==1.5.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
spacy==2.3.2
sqlparse==0.3.1
srsly==1.0.2
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
toml==0.10.1
torch==1.5.1+cpu
torchfile==0.1.0
torchvision==0.6.1+cpu
tornado==6.0.4
tqdm==4.48.0
traitlets==4.3.3
typed-ast==1.4.1
urllib3==1.25.9
visdom==0.1.8.9
wasabi==0.7.1
wcwidth==0.2.5
webdriver-manager==3.2.1
webencodings==0.5.1
websocket-client==0.57.0
whitenoise==5.1.0
wrapt==1.12.1
xlrd==1.2.0
xmltodict==0.12.0
yfinance==0.1.54
zipp==3.1.0
zope.interface==5.1.0
